Move Checker and Violation definitions to their own modules
