+++ src/eblint/checkers/violation.py
@@ -1,0 +1,6 @@
+import ast
+from typing import NamedTuple
+
+class Violation(NamedTuple):
+    node: ast.AST
+    message: str
